GCP/CloudFunctions/DownloadDataJMA/requirements.txt
functions-framework==3.*
requests
google-cloud-storage
google-cloud-logging
